+++ lib/cldr.ex
@@ -60,7 +60,7 @@ defmodule Cldr do
   ## Example
 
       iex> Cldr.version
-      {38, 0, 0}
+      {38, 1, 0}
 
   """
   @version Config.version()