+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Changelog for Cldr v2.20.0
+
+## Enhancements
+
+* Update to CLDR data version 39.
+
 # Changelog for Cldr v2.19.0
 
 This is the changelog for Cldr v2.19.0 released on February 6th.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-cldr/cldr/tags)